requirements.txt
streamlit
pinecone-client
pinecone-client[grpc]
PyPDF2
sentence-transformers